Fix format string for GC debug log message
Closes PR 13998

+++ racket/src/racket/src/thread.c
@@ -8676,7 +8676,7 @@ static void done_with_GC()
 }
 
 #ifdef MZ_PRECISE_GC
-static char *gc_num(char *nums, int v)
+static char *gc_num(char *nums, long long v)
 /* format a number with commas */
 {
   int i, j, len, clen, c, d;
@@ -8686,7 +8686,7 @@ static char *gc_num(char *nums, int v)
 
   v /= 1024; /* bytes => kbytes */
 
-  sprintf(nums+i, "%d", v);
+  sprintf(nums+i, "%lld", v);
   for (len = 0; nums[i+len]; len++) { }
   clen = len + ((len + ((nums[i] == '-') ? -2 : -1)) / 3);
